Broke out people page from being directly built into App.js
